add jupyter unit tests
