+++ setup.cfg
@@ -2,7 +2,7 @@
 source = qiskit_ibm_provider
 
 [mypy]
-exclude = test/integration/test_jupyter.py
+exclude = (test/integration/test_jupyter.py|test/unit/test_jupyter_dashboard.py)
 python_version = 3.7
 namespace_packages = True
 ignore_missing_imports = True
